Align net crate structure with content
