+++ net/Cargo.toml
@@ -5,9 +5,12 @@ edition = "2024"
 publish = false
 autobins = false
 
+[lib]
+path = "src/main.rs"
+
 [[bin]]
 name = "formal-web-net"
-path = "src/main.rs"
+path = "src/bin/net_process.rs"
 
 [dependencies]
 ipc-channel = "0.21"
